Update Linux battery module to properly parse ACPI support in the 2.6 kernel.
Note that a common kernel config option could cause trouble; if
'CONFIG_ACPI_DEBUG' is set, any call to the acpi battery status
interface will be VERY expensive, potentially as bad as freezing the
machine for 5-10 *seconds* at a time.  On my thinkpad T42p (about as
fast as they come here in early 2005), it was causing 200ms freezes.