Add my re-write of Jack Moffitt's thread library, used as "libicethread" in Icecast, among others. No copying and pasting was going on, but I did look at his version for ideas and guidance, and there are a bunch of striking similarities.
Since I am obviously wandering a grey area here, I emailed Jack about this
and he said it is okay.

He doesn't like the BSD license here very much, because of the many free-
riders that are already using Xiph.org's work without giving anything back --
he favors copyleft licenses. My own opinion in this regard is different and
biased like anyone else's, but I firmly believe that with these, BSD is the
way to go. They are so low-level that it is impossible to tell what they
might end up being used for (even within Xiph) and I do not want to force
people to libtoolize and separate these few lines of code just so they can
choose a license that is best for their project.

The API of this libthread version is still almost the same, just a little bit
more consistent in a few areas. The debugging facilities have been improved
and fleshed out, and there is no longer a dependency on libavl. My regression
test, using it within Icecast, worked out well, too.

TODO: Protect strerror().