Docbook to latex conversion of the spec by Max Horn.
We now build the specification document from latex source instead of the 
older docbook method since may developers have had trouble getting the 
docbook tools to work. It's also faster.

For now, the built documentation is still kept in svn and rebuilt only 
if --enable-docs is passed to configure. We may relax this to rebuilding
only if the tools are available once we're more confident it will work
with common TeX installations.