Add an awk script to build a list of exported symbols from MSVC .def files.
The Mach ld (Apple MacOS X) takes an -exported_symbols_list argument 
which, when passed a file containing a simple list of symbols, will
hide all symbols in a library which do not appear on the list. The
symbol names must be mangled.

This is, in fact, the simpler format, compared to the .def files used
my MSVC and the mingw32 compilers, or the Version script hack supported
by earlier versions of gcc. Since the MSVC build is the least flexible
in the kind of scripts is can run, it makes sense to maintain the symbol
list in .def format and derive the other two from it.

This script does that by parsing a useful subset of the .def file format 
and constructing the appropriate Mach ld formatted version.