Move the encoder and decoder source back into a single directory.
These were split when they were being worked on separately, and to 
simply building when they shared little code. However, for performance 
and reuse reasons, the encoder now shares significant code with the 
decoder and it doesn't make since to keep the source separate any more.

I've attempted to update all the build files, but the msvc build is 
untested and the xmingw32 build is only partially tested.

Also, add the prefix fedora 11 uses for the xmingw32 compiler.