Update the API documenation to the 0.9.14 release.
This carries some significant improvements in the fuction
descriptions over the earlier 0.9.8 version.

Built with doxygen 1.7.1 (which is older than the previous
commit's 1.7.4) because I was having trouble with the PDF
output from the 1.7.5 I have on my local machine. This was
produced on the jenkins vm, just like the pre-release docs.